Add file via upload
Baseball (MLB) Linear Regression model where 'target' variable is Runs Created.

Runs Created estimates a player's offensive contribution in terms of total runs. It combines a player's ability to get on base with his ability to hit for extra bases and is therefore a useful metric for evaluating/estimating the combined projected offense for a team or player.